auth: Split BasicAuth out into its own file
